+++ custom_components/edgeos/configuration/helpers/const.py
@@ -11,8 +11,6 @@ DOMAIN = "edgeos"
 DEFAULT_NAME = "EdgeOS"
 MANUFACTURER = "Ubiquiti"
 
-MAIN_VIEW = f"main_view_{DOMAIN}"
-
 DATA_KEYS = [
     CONF_HOST,
     CONF_USERNAME,